pyproject.toml
[project]
name = "pyauth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "phonenumberslite>=9.0.21",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
